+++ packages/connect-kit/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+
+- Close the ConnectWithLedgerLive modal when clicking outside.
+- Make the cursor a pointer for the "Install" link on the
+  ConnectWithLedgerLive modal.
+- Fix spacing around the "or" line on the ConnectWithLedgerLive modal
 
 ## 1.0.6 - 2022-12-13
 ### Changed